+++ Raytracing/MathHelper.h
@@ -5,18 +5,18 @@
 
 typedef struct Array2D
 {
-	float a;
-	float b;
+	double a;
+	double b;
 } Array2D;
 
 typedef struct Vector3D
 {
-	float x;
-	float y;
-	float z;
+	double x;
+	double y;
+	double z;
 } Vector3D;
 
-float DotProduct(Vector3D u, Vector3D v)
+double DotProduct(Vector3D u, Vector3D v)
 {
 	return u.x * v.x + u.y * v.y + u.z * v.z;
 }
@@ -31,9 +31,9 @@ Vector3D VectorSubstract(Vector3D v, Vector3D u)
 	return (Vector3D) { .x = v.x - u.x, .y = v.y - u.y, .z = v.z - u.z };
 }
 
-float Length(Vector3D u)
+double Length(Vector3D u)
 {
-	return sqrtf(u.x * u.x + u.y * u.y + u.z * u.z);
+	return sqrt(u.x * u.x + u.y * u.y + u.z * u.z);
 }
 
 Vector3D ScalarMul(float scalar, Vector3D u)
